Redirect to cancel_url from ThankYouView when order is not paid.
If a user successfully submits the CheckoutSelectionView a number of times
creating multiple orders. Then when they get to the thank you page refresh
the page a number of times, the other orders will also be marked as
complete. Now, of course, these orders will not have any payment records
attached but if the shop instance or shop admin doesn't check
Order.is_payed() they may send the product incorrectly.

Signed-off-by: Simon Luijk <simon@simonluijk.com>